full api cleanup where we expose only the necessary methods for creating commands and/or sub-commands
